```
name                                old time/op  new time/op  delta
WACDMEN-4                           2.20ms ± 1%  2.20ms ± 1%     ~     (p=0.690 n=5+5)
WACDMENdistance-4                   9.83ms ± 7%  1.74ms ± 1%  -82.32%  (p=0.008 n=5+5)
WACDME-4                             205ns ± 0%   205ns ± 0%     ~     (all equal)
WACDMEinv-4                          229ns ± 1%   226ns ± 0%   -1.14%  (p=0.000 n=5+4)
WACDMComovingDistance-4              357µs ± 2%   367µs ± 1%   +2.87%  (p=0.008 n=5+5)
WACDMComovingTransverseDistance-4    369µs ± 1%   356µs ± 1%   -3.39%  (p=0.008 n=5+5)
WACDMLuminosityDistance-4            393µs ± 1%   412µs ± 1%   +4.80%  (p=0.008 n=5+5)
WACDMLookbackTime-4                  370µs ± 1%   363µs ± 1%   -1.82%  (p=0.008 n=5+5)
WACDMNComovingDistance-4             3.55s ± 0%   3.62s ± 0%   +2.03%  (p=0.008 n=5+5)
WACDMNLuminosityDistance-4           3.58s ± 1%   3.54s ± 0%   -1.01%  (p=0.008 n=5+5)
WACDMNE-4                           10.1ms ± 1%   1.7ms ± 1%  -82.68%  (p=0.008 n=5+5)
WACDMComovingDistanceOM-4            947ns ± 2%   156ns ± 1%  -83.53%  (p=0.008 n=5+5)
WACDMLookbackTimeOM-4               1.33µs ± 4%  0.49µs ± 1%  -63.26%  (p=0.008 n=5+5)
WACDMComovingDistanceFlat-4         1.52µs ± 3%  0.66µs ± 1%  -56.85%  (p=0.008 n=5+5)
WACDMComovingDistancePositiveOk0-4   337µs ± 1%   331µs ± 1%   -1.73%  (p=0.008 n=5+5)
```